feat: bundle axios client
