--- packages/client-axios/README.md
@@ -1,101 +1,0 @@
-<div align="center">
-  <img alt="Hey API logo" height="150" src="https://heyapi.dev/images/logo-300w.png" width="150">
-  <h1 align="center"><b>Axios Client</b></h1>
-  <p align="center">🚀 Axios client for `@hey-api/openapi-ts` codegen.</p>
-</div>
-
-[Live demo](https://stackblitz.com/edit/hey-api-client-axios-example?file=openapi-ts.config.ts,src%2Fclient%2Fschemas.gen.ts,src%2Fclient%2Fsdk.gen.ts,src%2Fclient%2Ftypes.gen.ts,src%2FApp.tsx)
-
-## Features
-
-- seamless integration with `@hey-api/openapi-ts` ecosystem
-- type-safe response data and errors
-- response data validation and transformation
-- access to the original request and response
-- granular request and response customization options
-- minimal learning curve thanks to extending the underlying technology
-- support bundling inside the generated output
-- [platform](https://heyapi.dev/openapi-ts/integrations) for automating codegen builds
-
-## Dashboard
-
-Hey API is an ecosystem of products helping you build better APIs. Superpower your codegen and APIs with our platform.
-
-[Sign In](https://app.heyapi.dev) to Hey API platform.
-
-## Documentation
-
-Please visit our [website](https://heyapi.dev) for documentation, guides, migrating, and more.
-
-## Sponsors
-
-Love Hey API? Become our [sponsor](https://github.com/sponsors/hey-api).
-
-<h3 align="center">Gold</h3>
-
-<table align="center" style="justify-content: center;align-items: center;display: flex;">
-  <tbody>
-    <tr>
-      <td align="center">
-        <p></p>
-        <p>
-          <a href="https://kutt.it/pkEZyc" target="_blank">
-            <picture height="50px">
-              <source media="(prefers-color-scheme: dark)" srcset="https://heyapi.dev/images/stainless-logo-wordmark-480w.jpeg">
-              <img alt="Stainless logo" height="50px" src="https://heyapi.dev/images/stainless-logo-wordmark-480w.jpeg">
-            </picture>
-          </a>
-          <br/>
-          Generate best-in-class SDKs.
-          <br/>
-          <a href="https://kutt.it/pkEZyc" style="text-decoration:none;" target="_blank">
-            stainless.com
-          </a>
-        </p>
-        <p></p>
-      </td>
-    </tr>
-  </tbody>
-</table>
-
-<h3 align="center">Silver</h3>
-
-<table align="center" style="justify-content: center;align-items: center;display: flex;">
-  <tbody>
-    <tr>
-      <td align="center">
-        <a href="https://kutt.it/skQUVd" target="_blank">
-          <picture height="40px">
-            <source media="(prefers-color-scheme: dark)" srcset="https://heyapi.dev/images/scalar-logo-wordmark-480w.jpeg">
-            <img alt="Scalar logo" height="40px" src="https://heyapi.dev/images/scalar-logo-wordmark-480w.jpeg">
-          </picture>
-        </a>
-        <br/>
-        <a href="https://kutt.it/skQUVd" style="text-decoration:none;" target="_blank">
-          scalar.com
-        </a>
-      </td>
-    </tr>
-  </tbody>
-</table>
-
-<h3 align="center">Bronze</h3>
-
-<table align="center" style="justify-content: center;align-items: center;display: flex;">
-  <tbody>
-    <tr>
-      <td align="center">
-        <a href="https://kutt.it/YpaKsX" target="_blank">
-          <picture height="34px">
-            <source media="(prefers-color-scheme: dark)" srcset="https://heyapi.dev/images/kinde-logo-wordmark-dark-480w.webp">
-            <img alt="Kinde logo" height="34px" src="https://heyapi.dev/images/kinde-logo-wordmark-480w.jpeg">
-          </picture>
-        </a>
-      </td>
-    </tr>
-  </tbody>
-</table>
-
-## Migration Guides
-
-[OpenAPI Typescript Codegen](https://heyapi.dev/openapi-ts/migrating#openapi-typescript-codegen)
