--- packages/client-axios/package.json
@@ -1,78 +1,0 @@
-{
-  "name": "@hey-api/client-axios",
-  "version": "0.9.1",
-  "description": "🚀 Axios client for `@hey-api/openapi-ts` codegen.",
-  "homepage": "https://heyapi.dev/",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/hey-api/openapi-ts.git"
-  },
-  "bugs": {
-    "url": "https://github.com/hey-api/openapi-ts/issues"
-  },
-  "license": "MIT",
-  "author": {
-    "email": "lubos@heyapi.dev",
-    "name": "Hey API",
-    "url": "https://heyapi.dev"
-  },
-  "funding": "https://github.com/sponsors/hey-api",
-  "keywords": [
-    "axios",
-    "client",
-    "codegen",
-    "http",
-    "javascript",
-    "openapi",
-    "react",
-    "rest",
-    "svelte",
-    "swagger",
-    "typescript",
-    "vue"
-  ],
-  "type": "module",
-  "main": "./dist/index.cjs",
-  "module": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-  "exports": {
-    ".": {
-      "import": {
-        "types": "./dist/index.d.ts",
-        "default": "./dist/index.js"
-      },
-      "require": {
-        "types": "./dist/index.d.cts",
-        "default": "./dist/index.cjs"
-      }
-    },
-    "./package.json": "./package.json"
-  },
-  "sideEffects": false,
-  "files": [
-    "dist",
-    "LICENSE.md",
-    "src"
-  ],
-  "scripts": {
-    "build": "tsup && rollup -c && pnpm check-exports",
-    "check-exports": "attw --pack .",
-    "dev": "tsup --watch",
-    "prepublishOnly": "pnpm build",
-    "test:coverage": "vitest run --coverage",
-    "test:update": "vitest watch --update",
-    "test:watch": "vitest watch",
-    "test": "vitest run",
-    "typecheck": "vitest --typecheck --watch=false"
-  },
-  "peerDependencies": {
-    "@hey-api/openapi-ts": "< 2",
-    "axios": ">= 1.0.0 < 2"
-  },
-  "devDependencies": {
-    "@config/vite-base": "workspace:*",
-    "@hey-api/client-core": "workspace:*",
-    "@hey-api/openapi-ts": "workspace:*",
-    "axios": "1.8.2"
-  }
-}
