--- packages/client-axios/src/index.ts
@@ -1,20 +1,0 @@
-export { createClient } from './client';
-export type {
-  Client,
-  ClientOptions,
-  Config,
-  CreateClientConfig,
-  Options,
-  OptionsLegacyParser,
-  RequestOptions,
-  RequestResult,
-  TDataShape,
-} from './types';
-export { createConfig } from './utils';
-export type { Auth, QuerySerializerOptions } from '@hey-api/client-core';
-export {
-  buildClientParams,
-  formDataBodySerializer,
-  jsonBodySerializer,
-  urlSearchParamsBodySerializer,
-} from '@hey-api/client-core';
