--- packages/client-core/package.json
@@ -1,57 +1,0 @@
-{
-  "name": "@hey-api/client-core",
-  "version": "0.2.1",
-  "private": true,
-  "description": "Core utilities for type-safe @hey-api/openapi-ts clients",
-  "homepage": "https://heyapi.dev/",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/hey-api/openapi-ts.git"
-  },
-  "bugs": {
-    "url": "https://github.com/hey-api/openapi-ts/issues"
-  },
-  "license": "MIT",
-  "author": {
-    "email": "lubos@heyapi.dev",
-    "name": "Hey API",
-    "url": "https://heyapi.dev"
-  },
-  "type": "module",
-  "main": "./dist/index.cjs",
-  "module": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-  "exports": {
-    ".": {
-      "import": {
-        "types": "./dist/index.d.ts",
-        "default": "./dist/index.js"
-      },
-      "require": {
-        "types": "./dist/index.d.cts",
-        "default": "./dist/index.cjs"
-      }
-    },
-    "./package.json": "./package.json"
-  },
-  "sideEffects": false,
-  "files": [
-    "dist",
-    "LICENSE.md",
-    "src"
-  ],
-  "scripts": {
-    "build": "tsup && pnpm check-exports",
-    "check-exports": "attw --pack .",
-    "dev": "tsup --watch",
-    "prepublishOnly": "pnpm build",
-    "test:coverage": "vitest run --coverage",
-    "test:update": "vitest watch --update",
-    "test:watch": "vitest watch",
-    "test": "vitest run",
-    "typecheck": "vitest --typecheck --watch=false"
-  },
-  "devDependencies": {
-    "@config/vite-base": "workspace:*"
-  }
-}
