--- packages/client-core/tsconfig.base.json
@@ -1,14 +1,0 @@
-{
-  "compilerOptions": {
-    "declaration": true,
-    "esModuleInterop": true,
-    "module": "ESNext",
-    "moduleResolution": "Bundler",
-    "noImplicitOverride": true,
-    "noUncheckedIndexedAccess": true,
-    "noUnusedLocals": true,
-    "strict": true,
-    "target": "ES2022",
-    "useUnknownInCatchVariables": false
-  }
-}
