--- packages/client-core/tsconfig.json
@@ -1,7 +1,0 @@
-{
-  "extends": "./tsconfig.base.json",
-  "compilerOptions": {
-    "declaration": false,
-    "esModuleInterop": true
-  }
-}
