+++ packages/openapi-ts/src/plugins/@hey-api/client-axios/bundle/index.ts
@@ -1,0 +1,21 @@
+export type { Auth } from '../../client-core/bundle/auth';
+export type { QuerySerializerOptions } from '../../client-core/bundle/bodySerializer';
+export {
+  formDataBodySerializer,
+  jsonBodySerializer,
+  urlSearchParamsBodySerializer,
+} from '../../client-core/bundle/bodySerializer';
+export { buildClientParams } from '../../client-core/bundle/params';
+export { createClient } from './client';
+export type {
+  Client,
+  ClientOptions,
+  Config,
+  CreateClientConfig,
+  Options,
+  OptionsLegacyParser,
+  RequestOptions,
+  RequestResult,
+  TDataShape,
+} from './types';
+export { createConfig } from './utils';
