+++ packages/openapi-ts/src/plugins/@hey-api/client-fetch/__tests__/client.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 
-import { createClient } from '../client/client';
+import { createClient } from '../bundle/client';
 
 describe('buildUrl', () => {
   const client = createClient();
